requirements.txt
python-telegram-bot[job-queue]==21.8
pandas==2.2.0
python-dotenv==1.0.0
